[VisBuilder] Avoid prop passing by extracting custom hooks
- refactor meta field identification
