+++ src/plugins/vis_builder/public/application/utils/use/index.ts
@@ -4,6 +4,8 @@
  */
 
 export { useAggs } from './use_aggs';
-export { useVisualizationType } from './use_visualization_type';
 export { useIndexPatterns } from './use_index_pattern';
+export { useOnAddFilter } from './use_on_add_filter';
+export { useSampleHits } from './use_sample_hits';
 export { useSavedVisBuilderVis } from './use_saved_vis_builder_vis';
+export { useVisualizationType } from './use_visualization_type';
